Rename some components to be more accurate.
