+++ src/activities/imageRecognition/PredictionUpload.jsx
