+++ src/activities/imageRecognition/TrainingImageUpload.jsx
